+++ static/datavzrd.js
@@ -23,15 +23,15 @@ function renderTickPlot(ah, columns, title, slug_title, specs, is_float, precisi
         index += 1;
     }
     let row = 0;
-    let table_rows = $('#table').bootstrapTable('getData', {useCurrentPage: true});
+    let table_rows = $('#table').bootstrapTable('getData', { useCurrentPage: true });
     $(`table > tbody > tr td:nth-child(${index})`).each(
-        function() {
+        function () {
             var id = `${slug_title}-${row}`;
             this.classList.add("plotcell");
             const div = document.createElement("div");
             let value = table_rows[row][title];
             if (is_float && precision !== undefined) {
-                value = precision_formatter(precision,value);
+                value = precision_formatter(precision, value);
             }
             if (value != "") {
                 this.innerHTML = "";
@@ -43,7 +43,7 @@ function renderTickPlot(ah, columns, title, slug_title, specs, is_float, precisi
                 var s = specs;
                 s.data = {};
                 s.data.values = data;
-                var opt = {"actions": false};
+                var opt = { "actions": false };
                 vegaEmbed(div, JSON.parse(JSON.stringify(s)), opt);
             }
             row++;
@@ -57,15 +57,15 @@ function renderBarPlot(ah, columns, title, slug_title, specs, is_float, precisio
         index += 1;
     }
     let row = 0;
-    let table_rows = $('#table').bootstrapTable('getData', {useCurrentPage: true});
+    let table_rows = $('#table').bootstrapTable('getData', { useCurrentPage: true });
     $(`table > tbody > tr td:nth-child(${index})`).each(
-        function() {
+        function () {
             var id = `${slug_title}-${row}`;
             this.classList.add("plotcell");
             const div = document.createElement("div");
             let value = table_rows[row][title];
             if (is_float && precision !== undefined) {
-                value = precision_formatter(precision,value);
+                value = precision_formatter(precision, value);
             }
             if (value != "") {
                 this.innerHTML = "";
@@ -77,7 +77,7 @@ function renderBarPlot(ah, columns, title, slug_title, specs, is_float, precisio
                 var s = specs;
                 s.data = {};
                 s.data.values = data;
-                var opt = {"actions": false};
+                var opt = { "actions": false };
                 vegaEmbed(div, JSON.parse(JSON.stringify(s)), opt);
             }
             row++;
@@ -94,7 +94,7 @@ function renderDetailTickBarPlot(value, div, specs, title) {
         var s = specs;
         s.data = {};
         s.data.values = data;
-        var opt = {"actions": false};
+        var opt = { "actions": false };
         vegaEmbed(div, JSON.parse(JSON.stringify(s)), opt);
     }
 }
@@ -169,12 +169,24 @@ function shortenColumn(ah, columns, title, ellipsis, detail_mode, header_label_l
     }
     let row = 0;
     $(`table > tbody > tr td:nth-child(${index})`).each(
-        function() {
+        function () {
             value = this.innerHTML;
             if (value.length > ellipsis) {
                 this.innerHTML = `${value.substring(0, ellipsis)}<a tabindex="0" role="button" href="#" data-toggle="popover" data-trigger="focus" data-html='true' data-content='<div style="overflow: auto; max-height: 30vh; max-width: 25vw;">${value}</div>'>...</a>`;
             }
             row++;
+        }
+    );
+}
+
+function shortenHeaderRow(row, ellipsis, skip_label) {
+    $(`table > thead > tr:nth-child(${row + 1}) > td`).each(
+        function() {
+            value = this.innerHTML;
+            if (value.length > ellipsis && !skip_label) {
+                this.innerHTML = `${value.substring(0, ellipsis)}<a tabindex="0" role="button" href="#" data-toggle="popover" data-trigger="focus" data-html='true' data-content='<div style="overflow: auto; max-height: 30vh; max-width: 25vw;">${value}</div>'>...</a>`;
+            }
+            skip_label = false;
         }
     );
 }
@@ -187,7 +199,7 @@ function linkUrlColumn(ah, dp_columns, columns, title, link_url, detail_mode, he
     }
     let table_rows = $('#table').bootstrapTable('getData');
     $(`table > tbody > tr td:nth-child(${index})`).each(
-        function() {
+        function () {
             let row = this.parentElement.dataset.index;
             let value = table_rows[row][title];
             let link = link_url.replaceAll("{value}", value);